+++ extensions/git-branches-panel/src/treeProvider.ts
@@ -1,6 +1,11 @@
 import * as vscode from 'vscode';
 
-import { type BranchInfo, type BranchSortOrder, type BranchTreeNode } from './branchModel';
+import {
+  type BranchInfo,
+  type BranchSortOrder,
+  type BranchTreeNode,
+  type TreeBranch,
+} from './branchModel';
 import {
   fetchRemoteState,
   getBranches,
@@ -17,7 +22,12 @@ import {
   type BranchLoadOptions,
 } from './treeDataLoader';
 import { BranchTreeItem } from './treeItem';
-import { buildStatusBarText, buildStatusBarTooltipContent, findContainerNode } from './treePresentation';
+import {
+  buildStatusBarText,
+  buildStatusBarTooltipContent,
+  findContainerNode,
+  findDescendantBranches,
+} from './treePresentation';
 
 export { BranchTreeItem, type NodeType } from './treeItem';
 export type { BranchLoadOptions } from './treeDataLoader';
@@ -60,19 +70,21 @@ export class BranchTreeProvider implements vscode.TreeDataProvider<BranchTreeIte
       return this.nodesToItems(this.dataLoader.getTreeData());
     }
 
-    if ((element.nodeType !== 'folder' && element.nodeType !== 'section') || !element.containerPath) {
+    const containerKey = element.containerKey ?? element.containerPath;
+
+    if ((element.nodeType !== 'folder' && element.nodeType !== 'section') || !containerKey) {
       return [];
     }
 
     if (element.nodeType === 'section') {
-      const section = getBranchSectionKey(element.containerPath);
+      const section = getBranchSectionKey(element.containerPath ?? containerKey);
 
       if (section && !this.dataLoader.isSectionLoaded(section)) {
         await this.refresh({ sections: [section], fetchRemoteState: false });
       }
     }
 
-    const container = findContainerNode(this.dataLoader.getTreeData(), element.containerPath);
+    const container = findContainerNode(this.dataLoader.getTreeData(), containerKey);
     return container ? this.nodesToItems(container.children) : [];
   }
 
@@ -82,6 +94,10 @@ export class BranchTreeProvider implements vscode.TreeDataProvider<BranchTreeIte
 
   getCurrentBranch(): BranchInfo | undefined {
     return this.dataLoader.getCurrentBranch();
+  }
+
+  getDescendantBranches(containerKey: string): readonly TreeBranch[] {
+    return findDescendantBranches(this.dataLoader.getTreeData(), containerKey);
   }
 
   private nodesToItems(nodes: readonly BranchTreeNode[]): BranchTreeItem[] {